refactor: split distributed package into merge and worker sub-packages
distributed/ (4 orchestration files):
  DistributedScanExecutor, QueryAnalyzer, FilePartitioner, NodeDiscovery

distributed/merge/ (5 files):
  ResultMerger, AggregationReducer, TopKMerger, ResultSerializer, MergeStrategy

distributed/worker/ (6 files):
  WorkerQueryExecutor, WorkerCredentialResolver, WorkerQueryTransportAction,
  WorkerQueryAction, WorkerQueryRequest, WorkerQueryResponse
