+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/LakehousePlugin.java
@@ -21,8 +21,8 @@ import org.opensearch.lakehouse.action.LakehouseQueryTransportAction;
 import org.opensearch.lakehouse.action.LakehousePplRestAction;
 import org.opensearch.lakehouse.action.LakehouseSqlRestAction;
 import org.opensearch.lakehouse.distributed.NodeDiscovery;
-import org.opensearch.lakehouse.distributed.WorkerQueryAction;
-import org.opensearch.lakehouse.distributed.WorkerQueryTransportAction;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryAction;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryTransportAction;
 import org.opensearch.lakehouse.schema.IcebergSchemaContributor;
 import org.opensearch.action.ActionRequest;
 import org.opensearch.core.action.ActionResponse;
