+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/distributed/DistributedScanExecutor.java
@@ -18,6 +18,13 @@ import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.lakehouse.LakehousePlugin;
+import org.opensearch.lakehouse.distributed.merge.MergeStrategy;
+import org.opensearch.lakehouse.distributed.merge.ResultMerger;
+import org.opensearch.lakehouse.distributed.merge.ResultSerializer;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryAction;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryExecutor;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryRequest;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryResponse;
 import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.TransportException;
 import org.opensearch.transport.TransportResponseHandler;
