+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/distributed/merge/TopKMerger.java
@@ -6,7 +6,9 @@
  * compatible open source license.
  */
 
-package org.opensearch.lakehouse.distributed;
+package org.opensearch.lakehouse.distributed.merge;
+
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryResponse;
 
 import java.util.ArrayList;
 import java.util.Comparator;
