+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/distributed/merge/package-info.java
@@ -1,0 +1,15 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Result merging strategies for distributed query execution.
+ * <p>
+ * Contains the merge dispatcher, aggregation reducer, top-K merger,
+ * result serializer, and merge strategy enum.
+ */
+package org.opensearch.lakehouse.distributed.merge;
