+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/distributed/worker/WorkerQueryAction.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.lakehouse.distributed;
+package org.opensearch.lakehouse.distributed.worker;
 
 import org.opensearch.action.ActionType;
