+++ sandbox/plugins/lakehouse-iceberg/src/main/java/org/opensearch/lakehouse/distributed/worker/package-info.java
@@ -1,0 +1,15 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Worker-side execution for distributed queries.
+ * <p>
+ * Contains the worker query executor, credential resolver,
+ * transport action, and request/response DTOs.
+ */
+package org.opensearch.lakehouse.distributed.worker;
