+++ sandbox/plugins/lakehouse-iceberg/src/test/java/org/opensearch/lakehouse/distributed/merge/ResultMergerTests.java
@@ -6,9 +6,10 @@
  * compatible open source license.
  */
 
-package org.opensearch.lakehouse.distributed;
+package org.opensearch.lakehouse.distributed.merge;
 
 import org.apache.calcite.sql.SqlKind;
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryResponse;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.util.Comparator;
