+++ sandbox/plugins/lakehouse-iceberg/src/test/java/org/opensearch/lakehouse/distributed/merge/ResultSerializerTests.java
@@ -6,8 +6,9 @@
  * compatible open source license.
  */
 
-package org.opensearch.lakehouse.distributed;
+package org.opensearch.lakehouse.distributed.merge;
 
+import org.opensearch.lakehouse.distributed.worker.WorkerQueryResponse;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.util.ArrayList;
