+++ sandbox/plugins/lakehouse-iceberg/src/test/java/org/opensearch/lakehouse/distributed/worker/WorkerQueryTransportActionTests.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.lakehouse.distributed;
+package org.opensearch.lakehouse.distributed.worker;
 
 import org.opensearch.analytics.exec.DataWarehouseScanContext;
 import org.opensearch.analytics.exec.DataWarehouseQueryEngine;
